backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
reportlab==4.0.4
python-dotenv==1.0.0